// src/models/common/MetadataProperty.ts
export enum SongMetadataProperty {
  Title = 'TITLE',
  TitleEN = 'TITLEEN',
  TitleJA = 'TITLEJA',
  TitleCN = 'TITLECN',
  TitleTW = 'TITLETW',
  TitleKO = 'TITLEKO',
  Subtitle = 'SUBTITLE',
  SubtitleEN = 'SUBTITLEEN',
  SubtitleJA = 'SUBTITLEJA',
  SubtitleCN = 'SUBTITLECN',
  SubtitleTW = 'SUBTITLETW',
  SubtitleKO = 'SUBTITLEKO',
  BPM = 'BPM',
  Wave = 'WAVE',
  Offset = 'OFFSET',
  DemoStart = 'DEMOSTART',
  Genre = 'GENRE',
  ScoreMode = 'SCOREMODE',
  Maker = 'MAKER',
  Lyrics = 'LYRICS',
  SongVol = 'SONGVOL',
  SEVol = 'SEVOL',
  Side = 'SIDE',
  Life = 'LIFE',
  Game = 'GAME',
  HeadScroll = 'HEADSCROLL',
  BGImage = 'BGIMAGE',
  BGMovie = 'BGMOVIE',
  MovieOffset = 'MOVIEOFFSET',
}

export enum CourseMetadataProperty {
  Course = 'COURSE',
  Level = 'LEVEL',
  Balloon = 'BALLOON',
  ScoreInit = 'SCOREINIT',
  ScoreDiff = 'SCOREDIFF',
  BalloonNor = 'BALLOONNOR',
  BalloonExp = 'BALLOONEXP',
  BalloonMas = 'BALLOONMAS',
  Style = 'STYLE',
  Exam1 = 'EXAM1',
  Exam2 = 'EXAM2',
  Exam3 = 'EXAM3',
  GaugeIncr = 'GAUGEINCR',
  Total = 'TOTAL',
  HiddenBranch = 'HIDDENBRANCH',
}